first_phase/iris_prediction/routes/jwt_handler.py
from dotenv import load_dotenv
import datetime
import jwt
import os

load_dotenv()
SECRET_KEY = os.getenv("SECRET_KEY")

def create_jwt_token(user_id: str):
    payload = {
        "sub": user_id,
        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=2)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")

def decode_jwt_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None
